+++ QSopt/solver.c
@@ -256,7 +256,7 @@ static int parseargs (int ac,
 			readbasis = boptarg;
 			break;
 		case 'P':
-			precision = atoi (boptarg);
+			precision = (unsigned int)atoi (boptarg);
 			break;
 		case 'd':
 			simplexalgo = DUAL_SIMPLEX;
